Restructure teensy pulse detection tests
