+++ cpp/teensy/test/test_pulse_detection/test_pulse_detection.cpp
@@ -1,0 +1,8 @@
+#include "test_pulse_determination.h"
+
+
+int main(int argc, char** argv) {
+    UNITY_BEGIN();
+    RUN_TEST(pulse_determination::test_short_time_fourier_transform);
+    UNITY_END();
+}
